+++ Gemfile
@@ -32,6 +32,7 @@ gem 'pry-rails'
 gem 'bcrypt'
 gem 'sidekiq'
 gem 'activerecord-import'
+gem 'whenever'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
