+++ app/jobs/concerns/github_api.rb
@@ -1,0 +1,3 @@
+module GithubApi
+  extend ActiveSupport::Concern
+end
