+++ app/jobs/github/base_job.rb
@@ -1,0 +1,14 @@
+module Github
+  class BaseJob < ApplicationJob
+
+    private
+
+    def client
+      @client ||= ::Github::Client.get_client
+    end
+
+    def auth_client(token)
+      @auth_client ||= ::Github::Client.get_auth_client(token)
+    end
+  end
+end
