+++ app/jobs/github/sync_organization_repo_job.rb
@@ -1,0 +1,5 @@
+class SyncOrganizationRepoJob < BaseJob
+  queue_as :default
+
+  sidekiq_options retry: false
+end
