+++ app/views/dashboards/_repository.html.erb
@@ -1,0 +1,3 @@
+<div id="<% repository.name %>-tab">
+
+</div>
