+++ batches/sync_github_batch.rb
@@ -16,11 +16,11 @@ class SyncGithubBatch
   private
 
   def sync_github_repository(github_user, github_client)
-    SyncRepositoryJob.perform_later(github_user)
+    Github::SyncRepositoryJob.perform_later(github_user)
   end
 
-  def sync_github_organization(github_user, github_client)
-    SyncOrganizationJob.perform_now(github_user)
+  def sync_github_organization_repo(github_user, github_client)
+    Github::SyncOrganizationRepoJob.perform_now(github_user)
   end
 
   def init_client(github_user_token)
