+++ config/initializers/sidekiq.rb
@@ -1,6 +1,9 @@
 # NOTE: https://shohheyhey.com/docker-redis-sidekiq/
 
-redis_config = YAML.load_file('config/redis.yml')[Rails.env]
+redis_config = YAML.load_file('config/redis.yml')[ENV['RAILS_ENV']]
+p "@@@@@@@@@@@@@@@@@@@@@"
+p ENV['RAILS_ENV']
+p "@@@@@@@@@@@@@@@@@@@@@"
 redis_config['db'] = redis_config['db']['sidekiq']
 
 Sidekiq.configure_server do |config|
